+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/simple/SimpleListViewModel.kt
@@ -1,0 +1,8 @@
+package jp.androidbook.recyclerviewtemplate.simple
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class SimpleListViewModel @Inject constructor(): ViewModel()
