+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/simple/SimpleRecyclerFragment.kt
@@ -6,17 +6,22 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import dagger.hilt.android.AndroidEntryPoint
 import jp.androidbook.recyclerviewtemplate.OnTappedRecyclerViewListener
 import jp.androidbook.recyclerviewtemplate.databinding.FragmentSimpleRecyclerBinding
 
 /**
  * SimpleなRecyclerViewのサンプル
  */
+@AndroidEntryPoint
 class SimpleRecyclerFragment : Fragment(), OnTappedRecyclerViewListener {
 
     private lateinit var binding: FragmentSimpleRecyclerBinding
+
+    private val simpleListViewModel: SimpleListViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
