+++ app/src/main/java/jp/androidbook/recyclerviewtemplate/simple/SimpleRecyclerViewAdapter.kt
@@ -1,19 +1,17 @@
 package jp.androidbook.recyclerviewtemplate.simple
 
-import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import jp.androidbook.recyclerviewtemplate.OnTappedRecyclerViewListener
-import jp.androidbook.recyclerviewtemplate.databinding.ItemSimpleRecyclerViewBinding
 
 class SimpleRecyclerViewAdapter(
     private val arrayList: Array<String>,
     private val listener: OnTappedRecyclerViewListener
-) : RecyclerView.Adapter<SimpleRecyclerViewAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<SimpleListViewHolder>() {
 
     // ViewHolderを生成
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder.createViewHolder(parent)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SimpleListViewHolder {
+        return SimpleListViewHolder.createViewHolder(parent)
     }
 
     // RecyclerViewで表示するリストの数を返す
@@ -22,28 +20,11 @@ class SimpleRecyclerViewAdapter(
     }
 
     // ViewHolderのViewと値を繋げる
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.setText(arrayList[position])
+    override fun onBindViewHolder(holder: SimpleListViewHolder, position: Int) {
         holder.itemView.setOnClickListener {
             listener.onTapped(arrayList[position])
         }
-    }
 
-    class ViewHolder(
-        private val binding: ItemSimpleRecyclerViewBinding
-    ) : RecyclerView.ViewHolder(binding.root) {
-
-        fun setText(text: String) {
-            binding.textView.text = text
-        }
-
-        companion object {
-            fun createViewHolder(parent: ViewGroup): ViewHolder {
-                val layoutInflater = LayoutInflater.from(parent.context)
-                val binding = ItemSimpleRecyclerViewBinding.inflate(layoutInflater, parent, false)
-
-                return ViewHolder(binding)
-            }
-        }
+        holder.bind(arrayList[position])
     }
 }
